refactor: move models to different files
