+++ qualibrate_runner/config/models.py
@@ -2,7 +2,8 @@ from typing import Optional, Union
 
 from pydantic import BaseModel, ConfigDict
 
-from qualibrate_runner.core.models.last_run import LastRun, RunStatus
+from qualibrate_runner.core.models.enums import RunStatusEnum
+from qualibrate_runner.core.models.last_run import LastRun
 from qualibrate_runner.core.types import QGraphType, QNodeType
 
 
@@ -16,5 +17,5 @@ class State(BaseModel):
     def is_running(self) -> bool:
         return (
             self.last_run is not None
-            and self.last_run.status == RunStatus.RUNNING
+            and self.last_run.status == RunStatusEnum.RUNNING
         )
